Add option to list devices, fix handling of errors
The Tripp Lite USB interface is flaky. Sometimes it works, sometimes it
doesn't sometimes part of it works and others don't. So, if there is an
error getting a response just skip it. Users will need to make sure the
expected fields are actually present instead of assuming they are there
or have meaningful data.